renamed stylesheets directory to 'styles'
