+++ frontend/src/component/Header.tsx
@@ -1,4 +1,4 @@
-import '../stylesheets/Header.css'
+import '../styles/Header.css'
 
 export default function Header() {
     return <div id={"header"}>Placeholder</div>
