+++ frontend/src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './stylesheets/index.css';
+import './styles/index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
