+++ frontend/src/styles/EditIndexCard.css
